added locks

+++ gdrive/auth.py
@@ -1,9 +1,11 @@
 import os 
 from pydrive.auth import GoogleAuth
-
+from miscellaneous.locks import locks
 gauth = GoogleAuth()
 
 async def auth(client, message):
+    if not locks['auth']:
+        return
     FOLDER_MIME_TYPE = 'application/vnd.google-apps.folder'
     drive: GoogleDrive
     http = None
@@ -29,6 +31,8 @@ async def auth(client, message):
         await message.reply_text("Already AUTH")
 
 def revoke(client, message):
+    if not locks['revoke']:
+        return
     ID = str(message.chat.id)
     if os.path.isfile(ID):
         os.remove(ID)
@@ -37,6 +41,8 @@ def revoke(client, message):
         message.reply_text("Ha Ha Ha")
 
 async def token(client, message):
+    if not locks['token']:
+        return
     msg = message.text
     print(msg)
     ID = message.chat.id
